Add recursive search for .cpp and .cc files
