+++ src/creatures/players/imbuements/imbuements.h
@@ -21,6 +21,7 @@
 #define SRC_CREATURES_PLAYERS_IMBUEMENTS_IMBUEMENTS_H_
 
 #include "declarations.hpp"
+#include "creatures/players/imbuements/imbuement.h"
 #include "creatures/players/player.h"
 #include "utils/tools.h"
 
@@ -73,74 +74,6 @@ class Imbuements {
 		std::vector<CategoryImbuement> categoriesImbuement;
 
 		uint32_t runningid = 0;
-};
-
-class Imbuement
-{
-	public:
-		Imbuement(uint16_t initId, uint16_t initBaseId) :
-				id(initId), baseid(initBaseId) {}
-
-		uint16_t getID() const {
-			return id;
-		}
-
-		uint16_t getBaseID() const {
-			return baseid;
-		}
-
-		uint32_t getStorage() const
-		{
-			return storage;
-		}
-
-		bool isPremium() {
-			return premium;
-		}
-		std::string getName() const {
-			return name;
-		}
-		std::string getDescription() const {
-			return description;
-		}
-
-		std::string getSubGroup() const {
-			return subgroup;
-		}
-
-		uint16_t getCategory() const {
-			return category;
-		}
-
-		const std::vector<std::pair<uint16_t, uint16_t>>& getItems() const {
-			return items;
-		}
-
-		uint16_t getIconID() {
-			return icon + (baseid - 1);
-		}
-
-		uint16_t icon = 1;
-		int32_t stats[STAT_LAST + 1] = {};
-		int32_t skills[SKILL_LAST + 1] = {};
-		int32_t speed = 0;
-		uint32_t capacity = 0;
-		int16_t absorbPercent[COMBAT_COUNT] = {};
-		int16_t elementDamage = 0;
-
-		CombatType_t combatType = COMBAT_NONE;
-
-	protected:
-		friend class Imbuements;
-		friend class Item;
-
-	private:
-		bool premium = false;
-		uint32_t storage = 0;
-		uint16_t id, baseid, category = 0;
-		std::string name, description, subgroup = "";
-
-		std::vector<std::pair<uint16_t, uint16_t>> items;
 };
 
 #endif  // SRC_CREATURES_PLAYERS_IMBUEMENTS_IMBUEMENTS_H_
